cinemy/requirements.txt
ansicon==1.89.0
arrow==1.2.2
asgiref==3.5.2
backports.zoneinfo==0.2.1
black==22.3.0
blessed==1.19.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
Django==4.0.4
django-environ==0.8.1
django-nested-inline==0.4.5
django-picklefield==3.0.1
django-q==1.3.9
django-ratelimit==3.0.1
djangorestframework==3.13.1
flake8==4.0.1
idna==3.3
jinxed==1.2.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.1.1
platformdirs==2.5.2
pycodestyle==2.8.0
pyflakes==2.4.0
python-dateutil==2.8.2
pytz==2022.1
redis==3.5.3
requests==2.28.0
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
wcwidth==0.2.5
